Chore: re-enable "unused" linter
Re-enabling "unused" linter due it was disabled temporally to release a
new version.

The file test_env.go provides artifacts which are used by the tests but
the "unused" linter was taking into account as a regular src file given
the name is not ending with the suffix "_test" therefore it was not
consider as part of the tests so renaming it following that rule cause
that the linter disesteem it as part of the regular src.